refactor(group_store): extract membership tests into membership/tests.rs (#2306)
Phase 2/3 of #2306 (epic #2300). No behavior change — 65 test fns
and one membership-only inline helper (`nest_for_test`) move out of
the monolithic `group_store/tests.rs` into a focused
`group_store/membership/tests.rs`, organized next to the code they
exercise.

## Outcomes

  before:  tests.rs               9,116 LOC, 230 test fns
  after:   tests.rs               6,344 LOC, 150 test fns
           membership/tests.rs    2,431 LOC,  65 test fns

`cargo test -p calimero-context --lib` runs the same 315 tests as
before with the same outcomes (0 failed). Total LOC is preserved —
moved, not deleted.

## What moved

Tests grouped by their primary subject in the membership cluster:

- Core CRUD + lookup: `add_and_check_membership`, `remove_member`,
  `get_member_role`, `require_group_admin_rejects_non_admin`,
  `count_members_and_admins`, `list_members_with_offset_and_limit`,
  `has_direct_group_member_*`, `remove_group_member_clears_member_metadata`.
- Policy: `membership_policy_*`, `membership_view_*`.
- Path resolution / inheritance: `check_membership_*` (10),
  `enumerate_inherited_members_*` (6), `inherited_admin_walk_*`,
  `membership_path_inherited_admin_*`,
  `auth_and_crypto_walks_agree_at_max_namespace_depth_boundary`.
- Position-aware lookup: `membership_status_at_*` (7), `fast_path_*` (4).
- Effective capabilities + visibility derived from membership:
  `get_effective_member_capabilities_*` (7), `subgroup_visible_to_*` (5).
- Namespace anchors: `trusted_anchors_*` (7), `namespace_member_pubkeys_*` (3),
  `is_authoritative_namespace_identity_*`.

The membership-only inline helper `nest_for_test` moves alongside
(now in `test_fixtures.rs` — see below — since it would otherwise
break the non-membership tests that also use it).

## What stays in tests.rs

Anything whose primary subject is a different module: namespace
DAG / op-log / retry, group settings, context registration / tree,
group keys, signing keys, upgrades, capabilities, deny list,
metadata, op-application pipeline (`apply_local_*`), governance
broadcast, replica reconciliation, TEE policy lookup. These will be
extracted by future sub-issues of #2300 as those modules consolidate.

## Shared test fixtures

A new `crates/context/src/group_store/test_fixtures.rs` holds the
six helpers that both test files need:
- `test_store`, `test_group_id`, `test_meta` — basic setup
- `dummy_member_removed_op` — placeholder GroupOp for tests that
  don't exercise the convergence-detection path
- `sample_meta_with_admin` — variant of `test_meta` with a chosen
  admin pubkey
- `nest_for_test` — `nest_group(...).unwrap()` wrapper

Visibility is `pub(super)`: visible to all `group_store/` submodules,
invisible outside. Module is gated by `#[cfg(test)]` so the helpers
don't ship.

## Edge cases handled

- `super::namespace::*` paths inside moved tests rewritten to
  `super::super::namespace::*` (membership/tests.rs is one level
  deeper than tests.rs was).
- `super::membership::policy_rules::*` → `super::policy_rules::*`
  (now a same-module sibling within `membership/`).
- `super::is_open_chain_to_namespace` (which lives in
  `capabilities.rs`, not the membership cluster) → `super::super::`
  to reach back into `group_store/`.

## Verification

- 315 passed; 0 failed; 0 ignored — same as Phase 1
- `cargo build --workspace --lib` green
- `git diff --stat`: -2,611 lines in tests.rs, +2,431 lines in the
  new membership/tests.rs, +52 lines in test_fixtures.rs.
  Total LOC moved, not deleted.
